Fixed a bug that may create .php file on server

+++ discard.php
@@ -7,7 +7,14 @@ if (!isset($_COOKIE['hash'])) {
 	exit;
 }
 
-$filename = get_file_from_hash($_COOKIE['hash']);
+$hash = $_COOKIE['hash'];
+
+if (!preg_match('/^[a-z0-9]{32}$/', $hash)) {
+	setcookie('hash', gen_hash(), time() + 365 * 86400);
+	exit;
+}
+
+$filename = get_file_from_hash($hash);
 unlink($filename);
 rmdir(dirname($filename));
